WIP: Refactor location of reusable external command handler functionality
