+++ src/tooling/Elastic.Documentation.Tooling/Elastic.Documentation.Tooling.csproj
@@ -13,6 +13,7 @@
     <PackageReference Include="Github.Actions.Core" />
     <PackageReference Include="Crayon" />
     <PackageReference Include="Errata" />
+    <PackageReference Include="Proc"  />
   </ItemGroup>
 
   <ItemGroup>
