+++ src/tooling/docs-builder/Tracking/IRepositoryTracker.cs
@@ -1,0 +1,10 @@
+// Licensed to Elasticsearch B.V under one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information
+
+namespace Documentation.Builder.Tracking;
+
+public interface IRepositoryTracker
+{
+	IEnumerable<string> GetChangedFiles();
+}
